+++ apps/fluent-tester/src/E2E/TextExperimental/specs/ExperimentalText.spec.win.ts
@@ -6,7 +6,7 @@ import { TEXT_A11Y_ROLE, PAGE_TIMEOUT, BOOT_APP_TIMEOUT } from '../../common/con
 describe('Experimental Text Testing Initialization', function () {
   it('Wait for app load', () => {
     NavigateAppPage.waitForPageDisplayed(BOOT_APP_TIMEOUT);
-    expect(NavigateAppPage.isPageLoaded()).toBeTruthy();
+    expect(NavigateAppPage.isPageLoaded()).toBeTruthy(NavigateAppPage.ERRORMESSAGE_APPLOAD);
   });
 
   it('Click and navigate to Experimental Text test page', () => {
@@ -18,7 +18,8 @@ describe('Experimental Text Testing Initialization', function () {
     NavigateAppPage.clickAndGoToExperimentalTextPage();
     ExperimentalTextPageObject.waitForPageDisplayed(PAGE_TIMEOUT);
 
-    expect(ExperimentalTextPageObject.isPageLoaded()).toBeTruthy();
+    expect(ExperimentalTextPageObject.isPageLoaded()).toBeTruthy(ExperimentalTextPageObject.ERRORMESSAGE_PAGELOAD);
+    expect(ExperimentalTextPageObject.didAssertPopup()).toBeFalsy(ExperimentalTextPageObject.ERRORMESSAGE_ASSERT); // Ensure no asserts popped up
   });
 });
 
@@ -26,6 +27,8 @@ describe('Experimental Text Accessibility Testing', () => {
   it('Text - Validate accessibilityRole is correct', () => {
     ExperimentalTextPageObject.scrollToTestElement();
     ExperimentalTextPageObject.waitForPrimaryElementDisplayed(PAGE_TIMEOUT);
+
     expect(ExperimentalTextPageObject.getAccessibilityRole()).toEqual(TEXT_A11Y_ROLE);
+    expect(ExperimentalTextPageObject.didAssertPopup()).toBeFalsy(ExperimentalTextPageObject.ERRORMESSAGE_ASSERT);
   });
 });