+++ packages/components/ContextualMenu/src/Submenu.tsx
@@ -4,7 +4,7 @@ import { View, ScrollView, Platform } from 'react-native';
 import { submenuName, SubmenuProps, SubmenuSlotProps, SubmenuType, SubmenuRenderData, SubmenuState } from './Submenu.types';
 import { settings } from './Submenu.settings';
 import { IUseComposeStyling, compose } from '@uifabricshared/foundation-compose';
-import { useSelectedKey } from '@fluentui-react-native/interactive-hooks';
+import { IFocusable, useKeyDownProps, useSelectedKey } from '@fluentui-react-native/interactive-hooks';
 import { mergeSettings } from '@uifabricshared/foundation-settings';
 import { backgroundColorTokens, borderTokens } from '@fluentui-react-native/tokens';
 import { Callout } from '@fluentui-react-native/callout';
@@ -18,6 +18,21 @@ export const Submenu = compose<SubmenuType>({
   usePrepareProps: (userProps: SubmenuProps, useStyling: IUseComposeStyling<SubmenuType>) => {
     const { setShowMenu, maxWidth, maxHeight, shouldFocusOnMount = true, shouldFocusOnContainer = true, ...rest } = userProps;
 
+    /**
+     * On macOS, focus isn't placed by default on the first focusable element. We get around this by focusing on the inner FocusZone
+     * hosting the menu. For whatever reason, to get the timing _just_ right to actually focus, we need an additional `setTimeout`
+     *  on top of the `useLayoutEffect` hook.
+     */
+    const focusZoneRef = React.useRef<IFocusable>(null);
+
+    React.useLayoutEffect(() => {
+      if (Platform.OS === 'macos') {
+        setTimeout(() => {
+          focusZoneRef.current?.focus();
+        }, 0);
+      }
+    }, []);
+
     // Grabs the context information from ContextualMenu (onDismissMenu callback)
     const context = React.useContext(CMContext);
 
@@ -28,11 +43,13 @@ export const Submenu = compose<SubmenuType>({
       userProps?.onShow && userProps.onShow();
       context.isSubmenuOpen = true;
     }, [context]);
+
     const onDismiss = React.useCallback(() => {
       userProps?.onDismiss();
       setShowMenu(false);
       context.isSubmenuOpen = false;
     }, [context, setShowMenu]);
+
     const dismissCallback = React.useCallback(() => {
       onDismiss();
       context?.onDismissMenu();
@@ -55,6 +72,9 @@ export const Submenu = compose<SubmenuType>({
 
     const styleProps = useStyling(userProps, (override: string) => state[override] || userProps[override]);
 
+    // Explicitly override onKeyDown to override the native windows behavior of moving focus with arrow keys.
+    const onKeyDownProps = useKeyDownProps(onDismiss, 'ArrowLeft');
+
     const containerPropsWin32: IViewProps = {
       accessible: shouldFocusOnContainer,
       focusable: shouldFocusOnContainer && containerFocus,
@@ -69,10 +89,21 @@ export const Submenu = compose<SubmenuType>({
         onDismiss: onDismiss,
         setInitialFocus: shouldFocusOnMount,
       },
-      container: Platform.select({
-        macos: {},
-        default: containerPropsWin32,
-      }),
+      container: {
+        ...onKeyDownProps,
+        ...(Platform.OS === ('win32' as string) && { containerPropsWin32 }),
+      },
+      scrollView: {
+        contentContainerStyle: {
+          flexDirection: 'column',
+          flexGrow: 1,
+        },
+        showsVerticalScrollIndicator: true,
+      },
+      focusZone: {
+        componentRef: focusZoneRef,
+        focusZoneDirection: 'vertical',
+      },
     });
 
     return { slotProps, state };
@@ -81,6 +112,8 @@ export const Submenu = compose<SubmenuType>({
   slots: {
     root: Callout,
     container: View,
+    scrollView: ScrollView,
+    focusZone: FocusZone,
   },
   styles: {
     root: [backgroundColorTokens, borderTokens],
@@ -90,21 +123,32 @@ export const Submenu = compose<SubmenuType>({
     if (renderData.state == undefined) {
       return null;
     }
-    return (
-      <CMContext.Provider value={renderData.state.context}>
-        <Slots.root>
-          <Slots.container>
-            {Platform.OS === 'macos' ? (
-              <FocusZone focusZoneDirection={'vertical'}>{children}</FocusZone>
-            ) : (
-              <ScrollView contentContainerStyle={{ flexDirection: 'column', flexGrow: 1 }} showsVerticalScrollIndicator={true}>
-                {children}
-              </ScrollView>
-            )}
-          </Slots.container>
-        </Slots.root>
-      </CMContext.Provider>
-    );
+
+    // On macOS, wrap the children in a FocusZone to allow you to arrow-key through the menu items.
+    // Duplicating the JSX trees was the only way I could find to correctly render the optional slot.
+    if (Platform.OS === 'macos') {
+      return (
+        <CMContext.Provider value={renderData.state.context}>
+          <Slots.root>
+            <Slots.container>
+              <Slots.scrollView>
+                <Slots.focusZone>{children}</Slots.focusZone>
+              </Slots.scrollView>
+            </Slots.container>
+          </Slots.root>
+        </CMContext.Provider>
+      );
+    } else {
+      return (
+        <CMContext.Provider value={renderData.state.context}>
+          <Slots.root>
+            <Slots.container>
+              <Slots.scrollView>{children}</Slots.scrollView>
+            </Slots.container>
+          </Slots.root>
+        </CMContext.Provider>
+      );
+    }
   },
 });