+++ packages/utils/interactive-hooks/src/useKeyProps.win32.ts
@@ -37,12 +37,20 @@ export function getKeyCallbackWorker(userCallback?: KeyCallback, ...keys: string
 function getKeyUpPropsWorker(userCallback: KeyCallback, ...keys: string[]): KeyPressProps {
   return {
     onKeyUp: getKeyCallbackWorker(userCallback, ...keys),
+    // Optional on Win32 to know which key events to suppress natively
+    keyUpEvents: keys.map((keyCode) => {
+      return { key: keyCode };
+    }),
   };
 }
 
 function getKeyDownPropsWorker(userCallback: KeyCallback, ...keys: string[]): KeyPressProps {
   return {
     onKeyDown: getKeyCallbackWorker(userCallback, ...keys),
+    // Optional on Win32 to know which key events to suppress natively
+    keyDownEvents: keys.map((keyCode) => {
+      return { key: keyCode };
+    }),
   };
 }
 
@@ -50,7 +58,7 @@ function getKeyDownPropsWorker(userCallback: KeyCallback, ...keys: string[]): Ke
  * Re-usable hook for an onKeyUp event.
  * @param userCallback The function you want to be called once the key has been activated on key up
  * @param keys A string of the key you want to perform some action on. If undefined, always invokes userCallback
- * @returns KeyPressProps: An object containing the correct platform specific props to  handle key press
+ * @returns KeyPressProps: An object containing the correct platform specific props to handle key press
  */
 export const useKeyUpProps = memoize(getKeyUpPropsWorker);
 
@@ -66,6 +74,6 @@ export const useKeyDownProps = memoize(getKeyDownPropsWorker);
  * Re-usable hook for keyboard events. on macOS, this is onKeyDown, while on windows this is onKeyUp.
  * @param userCallback The function you want to be called once the key has been activated on key down
  * @param keys A string of the key you want to perform some action on. If undefined, always invokes userCallback
- * @returns KeyPressProps: An object containing the correct platform specific props to  handle key press
+ * @returns KeyPressProps: An object containing the correct platform specific props to handle key press
  */
 export const useKeyProps = memoize(getKeyUpPropsWorker);