+++ packages/components/Tabs/CHANGELOG.md
@@ -1,8 +1,39 @@
 # Change Log - @fluentui-react-native/tabs
 
-This log was last generated on Wed, 26 Jan 2022 20:31:04 GMT and should not be manually modified.
+This log was last generated on Wed, 09 Feb 2022 20:14:50 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.6.20
+
+Wed, 09 Feb 2022 20:14:50 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/focus-zone to v0.9.1
+- Bump @fluentui-react-native/interactive-hooks to v0.13.0
+- Bump @fluentui-react-native/pressable to v0.8.19
+
+## 0.6.19
+
+Fri, 04 Feb 2022 20:31:52 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/focus-zone to v0.9.0
+
+## 0.6.18
+
+Wed, 02 Feb 2022 02:29:07 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/focus-zone to v0.8.18
+- Bump @fluentui-react-native/icon to v0.11.6
+- Bump @fluentui-react-native/interactive-hooks to v0.12.2
+- Bump @fluentui-react-native/pressable to v0.8.18
+- Bump @fluentui-react-native/text to v0.11.9
+- Bump @uifabricshared/foundation-compose to v1.11.9
 
 ## 0.6.17