+++ packages/utils/interactive-hooks/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - @fluentui-react-native/interactive-hooks
 
-This log was last generated on Wed, 26 Jan 2022 20:31:04 GMT and should not be manually modified.
+This log was last generated on Wed, 09 Feb 2022 20:14:50 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.13.0
+
+Wed, 09 Feb 2022 20:14:50 GMT
+
+### Minor changes
+
+- Override keyUpEvents / keyDownEvents prop on win32 (sanajmi@microsoft.com)
+
+## 0.12.2
+
+Wed, 02 Feb 2022 02:29:07 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/icon to v0.11.6
 
 ## 0.12.1